Ajout des fonctions bloquantes de l'appartenance

+++ Model/Plateau.py
@@ -116,7 +116,8 @@ class Plateau():
         self.restart = False
         global counter
         counter = 1
-        self.overlayCounter = 0         
+        self.overlayCounter = 0   
+        self.property = 1      
 
     def save_game(self, filename):
         if filename.split(".")[-1] != "pickle":
@@ -376,17 +377,17 @@ class Plateau():
                 self.currentSpeed -= 10 
         self.soundMixer.playEffect("clickEffect")
 
-    def clearLand(self, grid_x1, grid_x2, grid_y1, grid_y2):
+    def clearLand(self, grid_x1, grid_x2, grid_y1, grid_y2, property):
         for xi in range(grid_x1, grid_x2+1):
                 for yi in range(grid_y1, grid_y2+1):
                         if self.map[xi][yi].sprite not in list_of_undestructible:
                             if self.map[xi][yi].sprite != "land":
                                 self.map[xi][yi].sprite = "land"
                                 self.map[xi][yi].indexSprite = randint(0, 57)
-                            if self.map[xi][yi].road :
+                            if self.map[xi][yi].road and self.map[xi][yi].road.property == property:
                                 self.map[xi][yi].road.delete()
                                 self.treasury = self.treasury - DESTRUCTION_COST
-                            if self.map[xi][yi].structure :
+                            if self.map[xi][yi].structure and self.map[xi][yi].structure.property == property :
                                 if self.map[xi][yi].structure.desc != "BurningBuilding" :
                                     self.map[xi][yi].structure.delete()
                                     self.treasury = self.treasury - DESTRUCTION_COST
@@ -394,41 +395,41 @@ class Plateau():
         self.collision_matrix = self.create_collision_matrix()
         self.foreground.initForegroundGrid()
 
-    def buildRoads(self, pattern, grid_x1, grid_x2, grid_y1, grid_y2):
+    def buildRoads(self, pattern, grid_x1, grid_x2, grid_y1, grid_y2, property):
         match(pattern):
             case 0:
                 for xi in range(grid_x1, grid_x2+1):
                     if self.map[xi][grid_y2].road == None and self.map[xi][grid_y2].structure == None and self.map[xi][grid_y2].sprite not in list_of_collision:
-                        Route(self.map[xi][grid_y2], self)
+                        Route(self.map[xi][grid_y2], self, property)
 
                 for yi in range(grid_y1, grid_y2+1):
                     if self.map[grid_x1][yi].road == None and self.map[grid_x1][yi].structure == None and self.map[grid_x1][yi].sprite not in list_of_collision:
-                        Route(self.map[grid_x1][yi], self)
+                        Route(self.map[grid_x1][yi], self, property)
             case 1:
                 for xi in range(grid_x1, grid_x2-1, -1):
                     if self.map[xi][grid_y1].road == None and self.map[xi][grid_y1].structure == None and self.map[xi][grid_y1].sprite not in list_of_collision:
-                        Route(self.map[xi][grid_y1], self)
+                        Route(self.map[xi][grid_y1], self, property)
                 for yi in range(grid_y1, grid_y2+1):
                     if self.map[grid_x2][yi].road == None and self.map[grid_x2][yi].structure == None and self.map[grid_x2][yi].sprite not in list_of_collision:
-                        Route(self.map[grid_x2][yi], self)
+                        Route(self.map[grid_x2][yi], self, property)
             case 2:
                 for xi in range(grid_x1, grid_x2+1):
                     if self.map[xi][grid_y1].road == None and self.map[xi][grid_y1].structure == None and self.map[xi][grid_y1].sprite not in list_of_collision:
-                        Route(self.map[xi][grid_y1], self)
+                        Route(self.map[xi][grid_y1], self, property)
                 for yi in range(grid_y1, grid_y2-1, -1):
                     if self.map[grid_x2][yi].road == None and self.map[grid_x2][yi].structure == None and self.map[grid_x2][yi].sprite not in list_of_collision:
-                        Route(self.map[grid_x2][yi], self)
+                        Route(self.map[grid_x2][yi], self, property)
             case 3:
                 for xi in range(grid_x1, grid_x2-1, -1):
                     if self.map[xi][grid_y2].road == None and self.map[xi][grid_y2].structure == None and self.map[xi][grid_y2].sprite not in list_of_collision:
-                        Route(self.map[xi][grid_y2], self)
+                        Route(self.map[xi][grid_y2], self, property)
                 for yi in range(grid_y1, grid_y2-1, -1):
                     if self.map[grid_x1][yi].road == None and self.map[grid_x1][yi].structure == None and self.map[grid_x1][yi].sprite not in list_of_collision:
-                        Route(self.map[grid_x1][yi], self)
+                        Route(self.map[grid_x1][yi], self, property)
         
         self.collision_matrix = self.create_collision_matrix()
 
-    def buildHousingSpot(self, grid_x1, grid_y1, grid_x2, grid_y2):
+    def buildHousingSpot(self, grid_x1, grid_y1, grid_x2, grid_y2, property):
         for xi in range(grid_x1, grid_x2+1):
             for yi in range(grid_y1, grid_y2+1):
                     for xcr in range (xi-2,xi+3,1) :
@@ -436,38 +437,38 @@ class Plateau():
                             if 0<=xcr<self.nbr_cell_x and 0<=ycr<self.nbr_cell_y:
                                 if not self.map[xi][yi].road and not self.map[xi][yi].structure and self.map[xi][yi].sprite not in list_of_collision and self.map[xi][yi].sprite not in list_of_undestructible:
                                     if self.map[xcr][ycr].road :
-                                        HousingSpot(self.map[xi][yi], self)
+                                        HousingSpot(self.map[xi][yi], self, property=property)
     
-    def buildPrefecture(self, grid_x1, grid_y1, grid_x2, grid_y2):
+    def buildPrefecture(self, grid_x1, grid_y1, grid_x2, grid_y2, property):
         for xi in range(grid_x1, grid_x2+1):
             for yi in range(grid_y1, grid_y2+1):
                 if self.map[xi][yi].getConnectedToRoad() > 0 :
                     if not self.map[xi][yi].road and not self.map[xi][yi].structure and self.map[xi][yi].sprite not in list_of_collision:
-                        Prefecture(self.map[xi][yi],self,(1,1),"Prefecture",1)
+                        Prefecture(self.map[xi][yi],self,(1,1),"Prefecture",1, property)
 
-    def buildEngineerPost(self, grid_x1, grid_y1, grid_x2, grid_y2):
+    def buildEngineerPost(self, grid_x1, grid_y1, grid_x2, grid_y2, property):
         for xi in range(grid_x1, grid_x2+1):
             for yi in range(grid_y1, grid_y2+1):
                 if self.map[xi][yi].getConnectedToRoad() > 0 :
                     if not self.map[xi][yi].road and not self.map[xi][yi].structure and self.map[xi][yi].sprite not in list_of_collision:
-                        EnginnerPost(self.map[xi][yi],self,(1,1),"EngineerPost",1)
+                        EnginnerPost(self.map[xi][yi],self,(1,1),"EngineerPost",1, property)
     
-    def buildWell(self, grid_x1, grid_y1, grid_x2, grid_y2):
+    def buildWell(self, grid_x1, grid_y1, grid_x2, grid_y2, property):
         for xi in range(grid_x1, grid_x2+1):
                 for yi in range(grid_y1, grid_y2+1):
                     if not self.map[xi][yi].road and not self.map[xi][yi].structure and self.map[xi][yi].sprite not in list_of_collision:
-                        Well(self.map[xi][yi],self,(1,1),"Well")
+                        Well(self.map[xi][yi],self,(1,1),"Well", property)
     
-    def buildSenate(self, grid_x1, grid_y1, grid_x2, grid_y2):
+    def buildSenate(self, grid_x1, grid_y1, grid_x2, grid_y2, property):
         for xi in range(grid_x1, grid_x2+1):
             for yi in range(grid_y1, grid_y2+1):
                 for xccl in range(xi, xi+5, 1) :
                     for yccl in range(yi, yi-5, -1 ) :
                         if self.map[xccl][yccl].road or self.map[xccl][yccl].structure or self.map[xccl][yccl].sprite in list_of_collision:
                             return
-        Senate(self.map[xi][yi],self,(5,5),"Senate")
+        Senate(self.map[xi][yi],self,(5,5),"Senate", property)
 
-    def buildFarm(self, grid_x1, grid_y1, grid_x2, grid_y2):
+    def buildFarm(self, grid_x1, grid_y1, grid_x2, grid_y2, property):
         #Vérifier que toutes les cases sont disponibles :
         for xi in range(grid_x1, grid_x2+1):
             for yi in range(grid_y1, grid_y2+1):
@@ -475,9 +476,9 @@ class Plateau():
                     for yccl in range(yi-1, yi+2, 1 ) :
                         if self.map[xccl][yccl].road or self.map[xccl][yccl].structure or self.map[xccl][yccl].sprite in list_of_collision:
                             return
-        WheatFarm(self.map[xi][yi],self,(2,2),"WheatFarm")
+        WheatFarm(self.map[xi][yi],self,(2,2),"WheatFarm", property)
             
-    def buildGranary(self, grid_x1, grid_y1, grid_x2, grid_y2):
+    def buildGranary(self, grid_x1, grid_y1, grid_x2, grid_y2, property):
         #Vérifier que toutes les cases sont disponibles :
         for xi in range(grid_x1, grid_x2+1):
             for yi in range(grid_y1, grid_y2+1):
@@ -485,18 +486,18 @@ class Plateau():
                     for yccl in range(yi, yi-3, -1 ) :
                         if self.map[xccl][yccl].road or self.map[xccl][yccl].structure or self.map[xccl][yccl].sprite in list_of_collision:
                             return
-        Granary(self.map[xi][yi],self,(3,3),"Granary")
+        Granary(self.map[xi][yi],self,(3,3),"Granary", property)
             
-    def buildMarket(self, grid_x1, grid_y1, grid_x2, grid_y2):
+    def buildMarket(self, grid_x1, grid_y1, grid_x2, grid_y2, property):
         for xi in range(grid_x1, grid_x2+1):
             for yi in range(grid_y1, grid_y2+1):
                 for xccl in range(xi, xi+2, 1) :
                     for yccl in range(yi, yi-2, -1 ) :
                         if self.map[xccl][yccl].road or self.map[xccl][yccl].structure or self.map[xccl][yccl].sprite in list_of_collision:
                             return
-        Market(self.map[xi][yi],self,(2,2),"Market")
+        Market(self.map[xi][yi],self,(2,2),"Market", property)
 
-    def buildTemple(self, grid_x1, grid_y1, grid_x2, grid_y2):
+    def buildTemple(self, grid_x1, grid_y1, grid_x2, grid_y2, property):
         #Vérifier que toutes les cases sont disponibles :
         for xi in range(grid_x1, grid_x2+1):
             for yi in range(grid_y1, grid_y2+1):
@@ -504,7 +505,7 @@ class Plateau():
                     for yccl in range(yi, yi-2, -1 ) :
                         if self.map[xccl][yccl].road or self.map[xccl][yccl].structure or self.map[xccl][yccl].sprite in list_of_collision:
                             return
-        Temple(self.map[xi][yi],self,(2,2),"Temple")
+        Temple(self.map[xi][yi],self,(2,2),"Temple", property)
 
     def update(self):
         if self.restart:
@@ -726,7 +727,9 @@ class Plateau():
         if self.roadWarning : self.screen.blit(self.road_warning_rectangle,(500,30))
 
         fpsText = self.minimalFont.render(f"FPS: {self.clock.get_fps():.0f}", 1, (255, 255, 255), (0, 0, 0))
+        propertyText = self.minimalFont.render(f"Player: {self.property}", 1, (255, 255, 255), (0, 0, 0))
         self.screen.blit(fpsText, (0, self.screen.get_height() - fpsText.get_height()))
+        self.screen.blit(propertyText, (0, self.screen.get_height() - fpsText.get_height() - propertyText.get_height()))
 
         # if state_control_panel=="reduced":
